Get basic layout and routing working for ui
This commit adds React Router (in Data Mode) to
the project.

The previous version (fp3) was build with React Router
in Framework mode, which it turns out is architecturally
problematic. It's a monolithic fullstack framework,
where the api and ui are tightly coupled and generated
(along with the code that lets them communicate) from
the same set of file.

This rewrite establishes the api and ui as clear and
totally separate entities, but uses data mode in the UI
for routing and data fetching from the API.