Add custom date scalar
The Date scalar in the graphql-scalars package has the annoying habit
of converting incoming values into JS Date objects with timezones
that shift the meaning of the date. This causes a bunch of issues with
incorrect dates and weird conversions to get it back to the string value
the database is expecting. The solution: Make our own.