Standardize on Deno
Initially the thinking was that we would hold back on deno until
we could figure out how to get the bundling setup working with the
esbuild-based `deno bundle` command. It turns out that deno will
cheerfully run rspack/rsbuild as is, so there isn't a tonne of value in
ripping out rspack and switching to esbuild.