add overload manager to address envoy warning
Previously, when starting the app via docker compose envoy would
print a warning to the console: "There is no configured limit to the
number of allowed active downstream connections. Configure a limit in
`envoy.resource_monitors.global_downstream_max_connections` resource
monitor." This commit changes the config to resolve this.