ui: Add SPA config for Caddy
This commit changes how we are using caddy to play nicer with our Single
Page Application (SPA) architecture.

We can now test caddy config changes locally with `cd ui && caddy run
--config Caddyfile`. The Caddyfile itself has been changed to fall
back to serving the index.html file in the case of paths it does not
recognize; a situation that occurs when clicking the verification link
which navigates to the path `/verify/somenumber`.